💄 Update the ui
