+++ static/app/views/organizationStats/index.tsx
@@ -395,16 +395,12 @@ export class OrganizationStats extends Component<Props> {
                 </ErrorBoundary>
               </div>
               <ErrorBoundary mini>
-                {isSingleProject && (
-                  <PanelHeading>
-                    <Title>{t('All Projects')}</Title>
-                  </PanelHeading>
-                )}
                 <UsageStatsProjects
                   organization={organization}
                   dataCategory={this.dataCategory}
                   dataCategoryName={this.dataCategoryName}
-                  projectIds={isSingleProject ? [ALL_ACCESS_PROJECTS] : this.projectIds}
+                  isSingleProject={isSingleProject}
+                  projectIds={this.projectIds}
                   dataDatetime={this.dataDatetime}
                   tableSort={this.tableSort}
                   tableQuery={this.tableQuery}
@@ -485,21 +481,6 @@ const Body = styled(Layout.Body)`
 const HeadingSubtitle = styled('p')`
   margin-top: ${space(0.5)};
   margin-bottom: 0;
-`;
-
-const Title = styled('div')`
-  font-weight: bold;
-  font-size: ${p => p.theme.fontSizeLarge};
-  color: ${p => p.theme.gray400};
-  display: flex;
-  flex: 1;
-  align-items: center;
-`;
-
-const PanelHeading = styled('div')`
-  display: flex;
-  margin-bottom: ${space(2)};
-  align-items: center;
 `;
 
 const PageControl = styled('div')`
